refactor(cli): extract harness selection helpers from ops-commands
Pre-existing ratchet debt: ops-commands sat above its complexity budget on
the integration baseline. The pure harness-selection helpers move to
ops-harness-selection.ts unchanged.
